Feat/sh 219 my study

+++ src/main/java/kr/co/studyhubinu/studyhubserver/comment/repository/CommentRepositoryCustom.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface CommentRepositoryCustom {
     List<CommentResponse> findSliceByPostIdWithUserId(Long postId, Long userId, Pageable pageable);
 
-    List<CommentResponse> findPreviewByPostId(Long postId, Long userId, Long commentPreviewCount);
+    List<CommentResponse> findPreviewByPostId(Long postId, Long userId, int commentPreviewCount);
 }
